--- Makefile
@@ -1,153 +1,0 @@
-include $(TOPDIR)/rules.mk
-
-PKG_NAME:=luci-app-xray
-PKG_VERSION:=1.29.0
-PKG_RELEASE:=1
-
-PKG_LICENSE:=MPLv2
-PKG_LICENSE_FILES:=LICENSE
-PKG_MAINTAINER:=yichya <mail@yichya.dev>
-PKG_BUILD_PARALLEL:=1
-
-include $(INCLUDE_DIR)/package.mk
-
-define Package/$(PKG_NAME)
-	SECTION:=Custom
-	CATEGORY:=Extra packages
-	TITLE:=LuCI Support for Xray
-	DEPENDS:=+luci-base +xray-core +dnsmasq +ca-bundle +PACKAGE_firewall4:kmod-nft-tproxy +PACKAGE_firewall:ipset +PACKAGE_firewall:iptables +PACKAGE_firewall:iptables-mod-tproxy
-	PKGARCH:=all
-endef
-
-define Package/$(PKG_NAME)/description
-	LuCI Support for Xray (Client-side Rendered).
-endef
-
-define Package/$(PKG_NAME)/config
-menu "luci-app-xray Configuration"
-	depends on PACKAGE_$(PKG_NAME)
-
-config PACKAGE_XRAY_INCLUDE_CLOUDFLARE_ORIGIN_ROOT_CA
-	bool "Include Cloudflare Origin Root CA"
-	default n
-
-config PACKAGE_XRAY_INFINITE_RETRY_ON_STARTUP
-	bool "Retry infinitely on Xray startup (may solve some startup problems)"
-	default n
-
-config PACKAGE_XRAY_RLIMIT_NOFILE_LARGE
-	bool "Increase Max Open Files Limit (recommended)"
-	default y
-
-config PACKAGE_XRAY_RESTART_DNSMASQ_ON_IFACE_CHANGE
-	bool "Restart dnsmasq on interface change (select this if using dnsmasq v2.87)"
-	default n
-
-config PACKAGE_XRAY_IGNORE_TP_SPEC_DEF_GW
-	bool "Ignore TP_SPEC_DEF_GW (select this if using private IPv4 address)"
-	default n
-
-choice
-	prompt "Limit memory use by setting rlimit_data (experimental)"
-	default PACKAGE_XRAY_RLIMIT_DATA_UNLIMITED
-	config PACKAGE_XRAY_RLIMIT_DATA_UNLIMITED
-		bool "Not limited"
-	config PACKAGE_XRAY_RLIMIT_DATA_SMALL
-		bool "Small limit (about 50MB)"
-	config PACKAGE_XRAY_RLIMIT_DATA_LARGE
-		bool "Large limit (about 321MB)"
-endchoice
-
-endmenu
-endef
-
-define Build/Compile
-endef
-
-define Package/$(PKG_NAME)/postinst
-#!/bin/sh
-if [[ ! -f $${IPKG_INSTROOT}/usr/share/xray/xray.init.replaced ]]; then
-	if [[ ! -f $${IPKG_INSTROOT}/etc/init.d/xray ]]; then
-	    echo "echo This file does nothing" > $${IPKG_INSTROOT}/etc/init.d/xray
-	fi
-	mv $${IPKG_INSTROOT}/etc/init.d/xray $${IPKG_INSTROOT}/usr/share/xray/xray.init.replaced
-	mkdir -p $${IPKG_INSTROOT}/etc/config
-	mv $${IPKG_INSTROOT}/tmp/xray.conf $${IPKG_INSTROOT}/etc/config/xray
-fi
-rm -f $${IPKG_INSTROOT}/tmp/xray.conf
-mkdir -p $${IPKG_INSTROOT}/etc/init.d
-mv $${IPKG_INSTROOT}/tmp/xray.init $${IPKG_INSTROOT}/etc/init.d/xray
-if [[ -z "$${IPKG_INSTROOT}" ]]; then
-	if [[ -f /etc/uci-defaults/xray ]]; then
-		( . /etc/uci-defaults/xray ) && rm -f /etc/uci-defaults/xray
-	fi
-	rm -rf /tmp/luci-indexcache* /tmp/luci-modulecache
-fi
-exit 0
-endef
-
-define Package/$(PKG_NAME)/conffiles
-/etc/config/xray
-endef
-
-define Package/$(PKG_NAME)/install
-	$(INSTALL_DIR) $(1)/tmp
-	$(INSTALL_BIN) ./root/etc/init.d/xray $(1)/tmp/xray.init
-	$(INSTALL_DATA) ./root/etc/config/xray $(1)/tmp/xray.conf
-	$(INSTALL_DIR) $(1)/etc/luci-uploads/xray
-	$(INSTALL_DIR) $(1)/etc/hotplug.d/iface
-	$(INSTALL_DIR) $(1)/etc/ssl/certs
-ifdef CONFIG_PACKAGE_XRAY_INCLUDE_CLOUDFLARE_ORIGIN_ROOT_CA
-	$(INSTALL_DATA) ./root/etc/ssl/certs/origin_ca_ecc_root.pem $(1)/etc/ssl/certs/origin_ca_ecc_root.pem
-endif
-	$(INSTALL_DIR) $(1)/etc/uci-defaults
-	$(INSTALL_BIN) ./root/etc/uci-defaults/xray $(1)/etc/uci-defaults/xray
-	$(INSTALL_DIR) $(1)/www/luci-static/resources/view
-	$(INSTALL_DATA) ./root/www/luci-static/resources/view/xray.js $(1)/www/luci-static/resources/view/xray.js
-	$(INSTALL_DIR) $(1)/usr/share/luci/menu.d
-	$(INSTALL_DATA) ./root/usr/share/luci/menu.d/luci-app-xray.json $(1)/usr/share/luci/menu.d/luci-app-xray.json
-	$(INSTALL_DIR) $(1)/usr/share/rpcd/acl.d
-	$(INSTALL_DATA) ./root/usr/share/rpcd/acl.d/luci-app-xray.json $(1)/usr/share/rpcd/acl.d/luci-app-xray.json
-	$(INSTALL_DIR) $(1)/usr/share/xray
-	$(LN) /var/run/xray.pid $(1)/usr/share/xray/xray.pid
-	$(LN) /usr/bin/xray $(1)/usr/share/xray/xray
-ifdef CONFIG_PACKAGE_XRAY_IGNORE_TP_SPEC_DEF_GW
-	$(INSTALL_DATA) ./root/usr/share/xray/ignore_tp_spec_def_gw $(1)/usr/share/xray/ignore_tp_spec_def_gw
-endif
-ifdef CONFIG_PACKAGE_XRAY_RESTART_DNSMASQ_ON_IFACE_CHANGE
-	$(INSTALL_DATA) ./root/usr/share/xray/restart_dnsmasq_on_iface_change $(1)/usr/share/xray/restart_dnsmasq_on_iface_change
-endif
-ifdef CONFIG_PACKAGE_XRAY_INFINITE_RETRY_ON_STARTUP
-	$(INSTALL_DATA) ./root/usr/share/xray/infinite_retry $(1)/usr/share/xray/infinite_retry
-endif
-ifdef CONFIG_PACKAGE_XRAY_RLIMIT_NOFILE_LARGE
-	$(INSTALL_DATA) ./root/usr/share/xray/rlimit_nofile_large $(1)/usr/share/xray/rlimit_nofile
-endif
-ifdef CONFIG_PACKAGE_XRAY_RLIMIT_DATA_SMALL
-	$(INSTALL_DATA) ./root/usr/share/xray/rlimit_data_small $(1)/usr/share/xray/rlimit_data
-endif
-ifdef CONFIG_PACKAGE_XRAY_RLIMIT_DATA_LARGE
-	$(INSTALL_DATA) ./root/usr/share/xray/rlimit_data_large $(1)/usr/share/xray/rlimit_data
-endif
-ifdef CONFIG_PACKAGE_firewall
-	$(INSTALL_BIN) ./root/etc/hotplug.d/iface/01-transparent-proxy-ipset.fw3 $(1)/etc/hotplug.d/iface/01-transparent-proxy-ipset
-	$(INSTALL_BIN) ./root/usr/share/xray/gen_ipset_rules.lua $(1)/usr/share/xray/gen_ipset_rules.lua
-	$(INSTALL_BIN) ./root/usr/share/xray/gen_ipset_rules_extra_normal.lua $(1)/usr/share/xray/gen_ipset_rules_extra.lua
-	$(INSTALL_BIN) ./root/usr/share/xray/firewall_include.lua $(1)/usr/share/xray/firewall_include.lua
-	$(INSTALL_DATA) ./root/usr/share/xray/init.fw3 $(1)/usr/share/xray/init.firewall
-	$(INSTALL_BIN) ./root/usr/share/xray/gen_config.lua $(1)/usr/share/xray/gen_config.lua
-endif
-ifdef CONFIG_PACKAGE_firewall4
-	$(INSTALL_BIN) ./root/etc/hotplug.d/iface/01-transparent-proxy-ipset.fw4 $(1)/etc/hotplug.d/iface/01-transparent-proxy-ipset
-	$(INSTALL_DATA) ./root/usr/share/xray/include.nft $(1)/usr/share/xray/include.nft
-	$(INSTALL_DIR) $(1)/etc/nftables.d
-	$(LN) /usr/share/xray/include.nft $(1)/etc/nftables.d/99-xray.nft
-	$(INSTALL_BIN) ./root/usr/share/xray/firewall_include.ut $(1)/usr/share/xray/firewall_include.ut
-	$(INSTALL_DATA) ./root/usr/share/xray/init.fw4 $(1)/usr/share/xray/init.firewall
-	$(INSTALL_BIN) ./root/usr/share/xray/gen_config.uc $(1)/usr/share/xray/gen_config.uc
-endif
-	$(INSTALL_DIR) $(1)/usr/libexec/rpcd
-	$(INSTALL_BIN) ./root/usr/libexec/rpcd/xray $(1)/usr/libexec/rpcd/xray
-endef
-
-$(eval $(call BuildPackage,$(PKG_NAME)))
