+++ fw3/root/etc/uci-defaults/xray_fw3
@@ -1,0 +1,16 @@
+#!/bin/sh
+uci get xray_fw3.@general[-1] >/dev/null 2>&1 || uci add xray_fw3 general >/dev/null 2>&1
+uci commit xray_fw3
+uci -q batch <<-EOF >/dev/null
+	delete ucitrack.@xray_fw3[-1]
+	add ucitrack xray_fw3
+	set ucitrack.@xray_fw3[-1].init=xray_fw3
+	commit ucitrack
+	delete firewall.xray_fw3
+	set firewall.xray_fw3=include
+	set firewall.xray_fw3.type=script
+	set firewall.xray_fw3.path=/var/etc/xray_fw3.include
+	set firewall.xray_fw3.reload=1
+	commit firewall
+EOF
+exit 0
