+++ fw3/root/usr/share/rpcd/acl.d/luci-app-xray.json
@@ -1,0 +1,15 @@
+{
+    "luci-app-xray": {
+        "description": "Grant access to xray configurations",
+        "read": {
+            "uci": [
+                "xray_fw3"
+            ]
+        },
+        "write": {
+            "uci": [
+                "xray_fw3"
+            ]
+        }
+    }
+}
