+++ fw4/root/etc/config/xray_fw4
@@ -1,0 +1,63 @@
+config general
+	option xray_bin '/usr/bin/xray'
+	option mark '255'
+	option tproxy_port_tcp '1080'
+	option tproxy_port_udp '1081'
+	option socks_port '1082'
+	option http_port '1083'
+	option dns_port '5300'
+	option dns_count '3'
+	option fast_dns '114.114.114.114'
+	option secure_dns '8.8.8.8'
+	option default_dns '1.1.1.1'
+	list bypassed_domain_rules 'geosite:cn'
+	list forwarded_domain_rules 'geosite:geolocation-!cn'
+	list blocked_domain_rules 'geosite:category-ads'
+	option transparent_proxy_enable '1'
+	option wan_bp_list '/dev/null'
+	option lan_target 'TP_SPEC_WAN_AC'
+	option lan_ifaces 'br-lan'
+	list wan_bp_ips '114.114.114.114'
+	option xray_api '1'
+	option main_server 'cfg024a8f'
+	option tproxy_udp_server 'cfg034a8f'
+	option tproxy_sniffing '1'
+	option routing_domain_strategy 'AsIs'
+	option conn_idle '300'
+	option loglevel 'warning'
+	option handshake '4'
+	option uplink_only '2'
+	option downlink_only '5'
+	option buffer_size '512'
+	option direct_bittorrent '1'
+
+config servers
+	option security 'auto'
+	option transport 'tcp'
+	option tcp_guise 'none'
+	option tls '0'
+	option tests_enabled 'none'
+	option protocol 'vless'
+	option server_port '443'
+	option password '00000000-0000-0000-0000-000000000000'
+	option vless_security 'none'
+	option vless_encryption 'none'
+	option server 'example.org'
+	option alias 'VLESS XTLS Splice Example'
+	option vless_flow 'xtls-rprx-splice'
+	option vless_tls 'xtls'
+	option vless_xtls_host 'example.org'
+	option vless_xtls_insecure '0'
+
+config servers
+	option password 'supersecret'
+	option transport 'tcp'
+	option tcp_guise 'none'
+	option server 'example.org'
+	option server_port '443'
+	option protocol 'trojan'
+	option alias 'Trojan Example'
+	option trojan_flow 'none'
+	option trojan_tls 'tls'
+	option trojan_tls_host 'example.org'
+	option trojan_tls_insecure '0'
