+++ fw4/root/etc/uci-defaults/xray_fw4
@@ -1,0 +1,10 @@
+#!/bin/sh
+uci get xray_fw4.@general[-1] >/dev/null 2>&1 || uci add xray_fw4 general >/dev/null 2>&1
+uci commit xray_fw4
+uci -q batch <<-EOF >/dev/null
+	delete ucitrack.@xray_fw4[-1]
+	add ucitrack xray_fw4
+	set ucitrack.@xray_fw4[-1].init=xray_fw4
+	commit ucitrack
+EOF
+exit 0
