--- root/usr/share/luci/menu.d/luci-app-xray.json
@@ -1,17 +1,0 @@
-{
-    "admin/services/xray": {
-        "title": "Xray",
-        "action": {
-            "type": "view",
-            "path": "xray"
-        },
-        "depends": {
-            "acl": [
-                "luci-app-xray"
-            ],
-            "uci": {
-                "xray": true
-            }
-        }
-    }
-}
