--- root/usr/share/rpcd/acl.d/luci-app-xray.json
@@ -1,15 +1,0 @@
-{
-    "luci-app-xray": {
-        "description": "Grant access to xray configurations",
-        "read": {
-            "uci": [
-                "xray"
-            ]
-        },
-        "write": {
-            "uci": [
-                "xray"
-            ]
-        }
-    }
-}
