--- root/usr/share/xray/init.fw3
@@ -1,31 +1,0 @@
-FIREWALL_INCLUDE="/usr/share/xray/firewall_include.lua"
-
-setup_firewall() {
-    logger -st xray[$$] -p4 "Setting ipset rules..."
-    lua /usr/share/xray/gen_ipset_rules.lua | ipset -! restore
-    /etc/hotplug.d/iface/01-transparent-proxy-ipset
-
-    logger -st xray[$$] -p4 "Generating firewall rules..."
-    /usr/bin/lua ${FIREWALL_INCLUDE} enable > $(uci get firewall.xray.path)
-
-    logger -st xray[$$] -p4 "Triggering firewall restart..."
-    /etc/init.d/firewall restart > /dev/null 2>&1
-}
-
-flush_firewall() {
-    logger -st xray[$$] -p4 "Flushing firewall rules..."
-    /usr/bin/lua ${FIREWALL_INCLUDE} flush > $(uci get firewall.xray.path)
-
-    logger -st xray[$$] -p4 "Triggering firewall restart..."
-    /etc/init.d/firewall restart > /dev/null 2>&1
-
-    logger -st xray[$$] -p4 "Flushing ipset rules..."
-    for setname in $(ipset -n list | grep "tp_spec"); do
-        ipset -! destroy $setname
-    done
-    /etc/hotplug.d/iface/01-transparent-proxy-ipset
-}
-
-impl_gen_config_file() {
-    /usr/bin/lua /usr/share/xray/gen_config.lua > /var/etc/xray/config.json
-}
