+++ shared/root/usr/share/xray/ignore_tp_spec_def_gw
